fix: resolve layout shifts by unifying data fetch on game page ss-057
