+++ src/libs/components/game-levels-preview/game-levels-preview.tsx
@@ -1,15 +1,10 @@
 import { EMPTY_ARRAY_LENGTH } from "~/libs/constants/constants";
 import { DataStatus } from "~/libs/enums/enums";
 import {
-	useAppDispatch,
 	useAppSelector,
-	useCallback,
-	useEffect,
-	useLanguageSync,
 	useTranslation,
 } from "~/libs/hooks/hooks";
 import { type GameDescriptionDto } from "~/libs/types/types";
-import { getLevelsList } from "~/modules/games/slices/actions";
 
 import { GameLevelsContent } from "./game-levels-content/game-levels-content";
 import styles from "./styles.module.css";
@@ -20,23 +15,9 @@ type Properties = {
 
 const GameLevelsPreview: React.FC<Properties> = ({ game }) => {
 	const { t } = useTranslation();
-	const dispatch = useAppDispatch();
 	const currentGameLevels = useAppSelector((state) => state.games.currentGameLevels);
 	const levelsStatus = useAppSelector((state) => state.games.levelsStatus);
 
-	const fetchLevels = useCallback(() => {
-		void dispatch(getLevelsList(game.id));
-	}, [dispatch, game.id]);
-
-	useLanguageSync(fetchLevels);
-
-	useEffect(() => {
-		if (levelsStatus === DataStatus.IDLE) {
-			fetchLevels();
-		}
-	}, [fetchLevels, levelsStatus]);
-
-	const isLoading = levelsStatus === DataStatus.PENDING || levelsStatus === DataStatus.IDLE;
 	const hasError = levelsStatus === DataStatus.REJECTED;
 	const hasLevels = Boolean(currentGameLevels && currentGameLevels.length > EMPTY_ARRAY_LENGTH);
 
@@ -50,7 +31,6 @@ const GameLevelsPreview: React.FC<Properties> = ({ game }) => {
 					game={game}
 					hasError={hasError}
 					hasLevels={hasLevels}
-					isLoading={isLoading}
 					levels={currentGameLevels ?? []}
 				/>
 			</section>
